requirements.txt
# Core Dependencies - Use compatible versions
langgraph>=0.2.0,<1.0.0
langchain>=0.3.0,<1.0.0
langchain-core>=0.3.0,<1.0.0
langchain-groq>=0.1.0,<1.0.0

# Vision & Image Processing
opencv-python-headless>=4.8.0  # Headless version (no GUI dependencies)
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0

# Vector DB & RAG
pinecone-client>=3.0.0
# weaviate-client>=4.0.0  # Alternative vector DB

# API & LLM
groq>=0.4.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Trading & Analysis
pandas>=2.0.0,<2.3.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web & API
requests>=2.31.0
beautifulsoup4>=4.12.0
streamlit>=1.30.0  # Interactive dashboard for demos
# selenium>=4.15.0  # Optional - only if needed for browser automation

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
loguru>=0.7.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Optional: News & Sentiment
# newsapi-python>=0.2.7
# transformers>=4.35.0  # For local sentiment analysis
